requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=2.121=mkl
blas-devel=3.9.0=21_win64_mkl
brotli-python=1.1.0=py38hd3f51b4_1
bzip2=1.0.8=hcfcfb64_5
ca-certificates=2024.2.2=h56e8100_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
colorama=0.4.6=pypi_0
cudatoolkit=11.5.2=h7b70026_13
flask=2.0.2=pypi_0
freetype=2.12.1=hdaf720e_2
ftfy=6.1.3=pypi_0
idna=3.6=pyhd8ed1ab_0
intel-openmp=2024.0.0=h57928b3_49841
itsdangerous=2.1.2=pypi_0
jinja2=3.1.3=pypi_0
jpeg=9e=hcfcfb64_3
lcms2=2.15=ha5c8aab_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=21_win64_mkl
libcblas=3.9.0=21_win64_mkl
libdeflate=1.17=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
libhwloc=2.9.3=default_haede6df_1009
libiconv=1.17=hcfcfb64_2
liblapack=3.9.0=21_win64_mkl
liblapacke=3.9.0=21_win64_mkl
libpng=1.6.42=h19919ed_0
libsqlite=3.44.2=hcfcfb64_0
libtiff=4.5.0=hf8721a0_2
libuv=1.44.2=hcfcfb64_1
libwebp-base=1.3.2=hcfcfb64_0
libxcb=1.13=hcd874cb_1004
libxml2=2.12.5=hc3477c8_0
libzlib=1.2.13=hcfcfb64_5
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.5=pypi_0
mkl=2024.0.0=h66d3029_49657
mkl-devel=2024.0.0=h57928b3_49657
mkl-include=2024.0.0=h66d3029_49657
msys2-conda-epoch=20160418=1
numpy=1.24.4=py38h1d91fd2_0
openjpeg=2.5.0=ha2aaf27_2
openssl=3.2.1=hcfcfb64_0
pillow=9.4.0=py38h087119c_1
pip=24.0=pyhd8ed1ab_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pysocks=1.7.1=pyh0701188_6
python=3.8.18=h4de0772_1_cpython
python_abi=3.8=4_cp38
pytorch=1.11.0=py3.8_cuda11.5_cudnn8_0
pytorch-mutex=1.0=cuda
regex=2023.12.25=pypi_0
requests=2.31.0=pyhd8ed1ab_0
setuptools=69.0.3=pyhd8ed1ab_0
tbb=2021.11.0=h91493d7_1
tk=8.6.13=h5226925_1
torchvision=0.12.0=py38_cu115
tqdm=4.66.1=pypi_0
typing_extensions=4.9.0=pyha770c72_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.2.0=pyhd8ed1ab_0
vc=14.3=hcf57466_18
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
wcwidth=0.2.13=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
zstd=1.5.5=h12be248_0
